requirements.txt
# Core Dependencies
Flask==3.0.0
Werkzeug==3.0.0
SQLAlchemy==2.0.22
# psycopg2-binary==2.9.9
gunicorn==21.2.0

# Flask Extensions
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
Flask-Limiter==3.5.0
Flask-Caching==2.1.0
Flask-Session==0.5.0
Flask-Cors==4.0.0
Flask-Admin==1.6.1
Flask-Talisman==1.1.0
Flask-Mail==0.9.1

# Security
PyJWT==2.8.0
cryptography==41.0.4
argon2-cffi==23.1.0
pyotp==2.9.0
passlib==1.7.4
bcrypt==4.0.1

# Monitoring & Logging
prometheus-flask-exporter==0.22.4
structlog==23.2.0
python-json-logger==2.0.7
sentry-sdk==1.31.0
psutil==5.9.5

# Testing
pytest==7.4.2
pytest-flask==1.3.0
pytest-cov==4.1.0
pytest-mock==3.11.1
pytest-asyncio==0.21.1
pytest-timeout==2.1.0
pytest-xdist==3.3.1
factory-boy==3.3.0

# Development
python-dotenv==1.0.0
black==23.9.1
flake8==6.1.0
mypy==1.5.1
pre-commit==3.4.0
isort==5.12.0
pylint==3.0.1
bandit==1.7.5

# Type Stubs
types-PyJWT>=0.1.3
types-redis==4.6.0.5
types-Flask==1.1.6
types-SQLAlchemy==1.4.53.38
types-requests==2.31.0.2
# types-pytest==7.4.0.0

# Cache & Queue
redis==5.0.1
celery==5.3.4
kombu==5.3.2

# Documentation
Sphinx==7.2.6
sphinx-rtd-theme==1.3.0
sphinx-autodoc-typehints==1.24.0
myst-parser==2.0.0